First cut at gate thunks.
Gate thunks are a special class of thunks that mediate access to a private
data structure.
A different gate class must be created for each data type to wrap.
For each gate class, we reserve a special piece of address space that is
guaranteed to not be mapped. This is the thunk token space and has the
same size as the data type wrapped by the gate class.
The user can freely manipulate and delegate thunk token space capabilities,
that authorise access to their respective sub-object range in the wrapped
data type.
The gate thunk can then be invoked with a thunk token: it will project
the token bounds and permission onto the wrapped data and return a
properly restricted pointer to the underlying data.

Additional fixes and supporting modifications.
Rework thunk template patching to allow static definition of metaclasses.